% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_grab.R
\name{box_stats}
\alias{box_stats}
\title{Box Stats Function}
\usage{
box_stats(team_num, year, type, game_count = NULL, bothteams = TRUE)
}
\arguments{
\item{team_num}{The number of the team. You can use the baseballr package to look up team_num.}

\item{year}{The year of the season you want.}

\item{type}{hitting, pitching, or fielding depending on what stats you want.}

\item{bothteams}{By default, bothteams is true and stats for your selected team and each opponent will be included. To only get the team you selected set to false.}
}
\description{
This function returns a data.frame of box stats with each player and game for the entire season
}
\examples{
box_stats(457,2018,'hitting')
box_stats()
}
\keyword{baseball,}
\keyword{college}
\keyword{ncaa,}
