% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{by_team}
\alias{by_team}
\title{by_team function}
\usage{
by_team(dframe, type = "sum")
}
\arguments{
\item{dframe}{the dataframe returned in advanced_stats or box_stats}

\item{type}{Options are sum, mean, max, min}
}
\description{
This function takes a dataframe from advanced_stats or box_stats and aggregates.
}
\examples{
by_team(df,'sum')
by_team()
}
\keyword{baseball,}
\keyword{college}
\keyword{ncaa,}
