% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo.R
\name{elo}
\alias{elo}
\title{ELO Function}
\usage{
elo(start_date, end_date, start_score = 1000, k_num = 10)
}
\arguments{
\item{start_date}{The date you want to start from the start_score.}

\item{end_date}{When you would like to end}

\item{k_num}{K value you would like to use for results.}

\item{start_sore}{What score you would like for teams to begin at}
}
\description{
This function returns a data.frame of elo for each game for every team from start date to end date. Earliest available
data is February of 2014.
}
\examples{
elo('02/10/2014','09/22/2018')
box_stats()
}
\keyword{baseball,}
\keyword{college,}
\keyword{elo}
\keyword{ncaa,}
