% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webpage_codes.R
\name{game_codes}
\alias{game_codes}
\title{game_codes function}
\usage{
game_codes(team_num, Year)
}
\arguments{
\item{team_num}{The team number to grab codes for}

\item{year}{The year to grab codes for}
}
\description{
This function takes a dataframe from advanced_stats or box_stats and aggregates.
}
\keyword{baseball,}
\keyword{college}
\keyword{game_codes()}
\keyword{ncaa,}
