% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_grab.R
\name{game_stats}
\alias{game_stats}
\title{Game Stats Function}
\usage{
game_stats(team_num, year, game_count = NULL, bothteams = TRUE)
}
\arguments{
\item{team_num}{The number of the team. You can use the baseballr package to look up team_num.}

\item{year}{The year of the season you want.}

\item{game_count}{The number of games back from latest point in pull. For example, 5 would give you last 5 games of the season. If null, the entire season is grabbed.}

\item{bothteams}{By default, bothteams is true and stats for your selected team and each opponent will be included. To only get the team you selected set to false.}
}
\description{
This function returns a data.frame of game statistics for a team in a given season
}
\examples{
game_stats(457,2018)
game_stats()
}
\keyword{baseball,}
\keyword{college}
\keyword{ncaa,}
