% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_team_dropdown.R
\name{get_ncaa_baseball_teams}
\alias{get_ncaa_baseball_teams}
\title{Get NCAA Baseball Teams}
\usage{
get_ncaa_baseball_teams()
}
\value{
A tibble with columns \code{org_id} and \code{school_name}
}
\description{
Scrapes the org_id and school name for all institutions sponsoring baseball.
}
