% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_team_roster.R
\name{get_roster}
\alias{get_roster}
\title{Get Roster for a Team-Season}
\usage{
get_roster(season_id)
}
\arguments{
\item{season_id}{Integer season ID}
}
\value{
A tibble with one row per player and columns:
\itemize{
\item number, name, player_url, class, position, height, bats, throws, hometown, highschool
}
}
\description{
Scrapes the roster table for a given team-season.
}
