% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webpage_codes.R
\name{ncaaYearCodes}
\alias{ncaaYearCodes}
\title{ncaaYearCodes function}
\usage{
ncaaYearCodes(year)
}
\arguments{
\item{year}{The year to grab codes for}
}
\description{
This function takes a dataframe from advanced_stats or box_stats and aggregates.
}
\keyword{baseball,}
\keyword{college}
\keyword{ncaa,}
\keyword{ncaaYearCodes()}
