% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_school_codes.R
\docType{data}
\name{school_codes}
\alias{school_codes}
\title{NCAA School Codes}
\format{
A tibble with columns:
\describe{
\item{org_id}{integer, unique team ID}
\item{school_name}{character, institution name}
}
}
\source{
\url{https://stats.ncaa.org/teams/history}
}
\usage{
school_codes
}
\description{
A cached dataset of \code{org_id} and \code{school_name} scraped from stats.ncaa.org.
}
\keyword{datasets}
