% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_lookup.R
\name{team_lookup}
\alias{team_lookup}
\title{Team Lookup Function
#'
This function returns the team numbers based on school name. These numbers are needed for other functions. Wraps baseballr function.}
\usage{
team_lookup(teams, year)
}
\arguments{
\item{teams}{A vector of teams that you would like returned}

\item{year}{The year of the season you want. Generally, team numbers stay consistent through multiple years.}
}
\description{
Team Lookup Function
#'
This function returns the team numbers based on school name. These numbers are needed for other functions. Wraps baseballr function.
}
\examples{
team_lookup(c('Florida','Coastal Caro.','Virginia'),2018)
box_stats()
}
\keyword{baseball,}
\keyword{college}
\keyword{ncaa,}
